requirements.txt
fastapi~=0.111.0
uvicorn~=0.27.0
pyyaml~=6.0.1
httpx[socks]~=0.28.0
brotlicffi~=1.1.0.0
